+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.7] / 2021-12-17
+- Add Icon/Image/License on IssRevitBuilder
 - Remove `.addin` Example
 - Set IHazMainProject on IHazPackageBuilderProject `Name`